Filter out artificial click events
I encountered this when working on a form that you can submit by hitting "enter". It seems to trigger an artificial click event on the form's submit button (see here: http://stackoverflow.com/questions/11760030/click-event-on-button-is-trigged-when-submitting-a-form-with-enter). The only difference I could see between this event and an actual click event is that the pageX and pageY on the artificial one were 0. I don't know if there are other circumstances in which these click events would be triggered, but I think ignoring them is the less surprising behavior.