Fixes issues #37 #29 #26 and #25
Seems like a lot of people were running into the same issue as me, where the filter wouldn't work for nodes added to the DOM after the element with off-click rendered. The solution might not be the best since we're having to run a querySelector every time there's a click but I think the benefits outweigh the cons.